Move convertXmlToBlockly tests to correct test filepath
