--- apps/test/unit/templates/convertXmlToBlockly.js
@@ -1,46 +1,0 @@
-/** @file Tests for convertXmlToBlockly utility */
-import {expect} from '../../util/reconfiguredChai';
-import setupBlocklyGlobal from '../../util/setupBlocklyGlobal';
-import {convertXmlToBlockly} from '@cdo/apps/templates/instructions/utils';
-
-describe('convertXmlToBlockly', function() {
-  beforeEach(() => {
-    // Make sure Blockly is setup in the global namespace because convertXmlToBlockly
-    // depends on it.
-    setupBlocklyGlobal();
-  });
-
-  it("does nothing if there's no xml", function() {
-    const container = document.createElement('div');
-    const content = '<p>Some random content</p>';
-    container.innerHTML = content;
-    convertXmlToBlockly(container);
-    expect(container.innerHTML).to.equal(content);
-    expect(container.getElementsByTagName('svg').length).to.equal(0);
-  });
-
-  it('creates a block space if it detects xml', function() {
-    const container = document.createElement('div');
-    const content =
-      '<p>Some random content</p><xml><block type="variables_get"><title name="VAR">i</title></block></xml>';
-    container.innerHTML = content;
-    expect(container.innerHTML).to.equal(content);
-    expect(container.getElementsByTagName('svg').length).to.equal(0);
-    convertXmlToBlockly(container);
-    expect(container.innerHTML).to.not.equal(content);
-    expect(container.getElementsByTagName('svg').length).to.equal(2);
-  });
-
-  it('ignores comments in xml', function() {
-    // For example, those generated by React
-    const container = document.createElement('div');
-    const content =
-      '<xml><!-- react-text: 1 --><block type="variables_get"><!-- react-text: 2 --><title name="VAR">i</title></block></xml>';
-    container.innerHTML = content;
-    expect(container.innerHTML).to.equal(content);
-    expect(container.getElementsByTagName('svg').length).to.equal(0);
-    convertXmlToBlockly(container);
-    expect(container.innerHTML).to.not.equal(content);
-    expect(container.getElementsByTagName('svg').length).to.equal(2);
-  });
-});
